// packages/v-component/src/components/q-dynamic-tabs/setting/types.ts
export interface ITabsData {
  id: string;
  title: string;
  icon: string;
  bgImage: string;
  hoverImage: string;
  selectedImage: string;
  bgColor: string;
  hoverBgColor: string;
  selectedBgColor: string;
  color: string;
  hoverColor: string;
  selectedColor: string;
  closeable: boolean;
}